refactor : inner class StationResponse 별도로 분리
